requirements.txt
annotated-doc==0.0.4
anyio==4.12.1
catboost==1.2.8
certifi==2026.1.4
click==8.3.1
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
filelock==3.20.3
fonttools==4.61.1
fsspec==2026.2.0
graphviz==0.21
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.4.1
idna==3.11
Jinja2==3.1.6
joblib==1.5.3
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.8
mpmath==1.3.0
narwhals==2.16.0
networkx==3.6.1
numpy==1.26.4
packaging==26.0
pandas==3.0.0
pillow==12.1.0
plotly==6.5.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
PyYAML==6.0.3
rdkit-pypi==2022.9.5
regex==2026.1.15
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.0
shellingham==1.5.4
six==1.17.0
sympy==1.13.1
threadpoolctl==3.6.0
tokenizers==0.22.2
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchvision==0.20.1+cu121
tqdm==4.67.3
transformers==5.1.0
typer-slim==0.21.2
typing_extensions==4.15.0
tzdata==2025.3
